Changes mayor_id on Leaderboard to int

+++ boundlexx/boundless/models.py
@@ -378,7 +378,7 @@ class LeaderboardRecord(models.Model):
     world_poll = models.ForeignKey("WorldPoll", on_delete=models.CASCADE)
     world_rank = models.PositiveSmallIntegerField(_("World Rank"))
     guild_tag = models.CharField(_("Guild Tag"), max_length=7)
-    mayor_id = models.PositiveSmallIntegerField()
+    mayor_id = models.PositiveIntegerField()
     mayor_name = models.CharField(max_length=64)
     mayor_type = models.PositiveSmallIntegerField()
     name = models.CharField(max_length=64)
